extract and test logout
